// src/redux/user/users.types.ts
export enum IUsersTypes {
  //Get user on app open
  START_GET_USER = 'START_GET_USER',
  SUCCESS_USER_GET = 'SUCCESS_USER_GET',
  FAILURE_USER_GET = 'FAILURE_USER_GET',
  //Login Types
  START_LOGIN_USER = 'START_LOGIN_USER',
  START_LOGIN_PROVIDER = 'START_LOGIN_PROVIDER',
  SUCCESS_LOGIN = 'SUCCESS_LOGIN',
  FAILURE_LOGIN = 'FAILURE_LOGIN',
  //SignupTypes
  START_SIGNUP_USER = 'START_SIGNUP_USER',
  START_SIGNUP_PROVIDER = 'START_SIGNUP_PROVIDER',
  SUCCESS_SIGNUP = 'SUCCESS_SIGNUP',
  FAILURE_SIGNUP = 'FAILURE_SIGNUP',
  // clean errors
  CLEAN_ERRORS = 'CLEAN_ERRORS',
  //reset password
  START_RESET_PASSWORD = 'START_RESET_PASSWORD',
  SUCCESS_RESET_PASSWORD = 'SUCCESS_RESET_PASSWORD',
  FAILURE_RESET_PASSWORD = 'FAILURE_RESET_PASSWORD',
  // update me
  START_UPDATE_ME = 'START_UPDATE_ME',
  UPDATE_ME_SUCCESS = 'UPDATE_ME_SUCCESS',
  UPDATE_ME_FAILURE = 'UPDATE_ME_FAILURE',
  // change passsword
  START_CHANGE_PASSWORD = 'START_CHANGE_PASSWORD',
  CHANGE_PASSWORD_SUCCESS = 'CHANGE_PASSWORD _SUCCESS',
  CHANGE_PASSWORD_FAILURE = 'CHANGE_PASSWORD _FAILURE',
  // upload pdf
  UPLOAD_START = 'UPLOAD_START',
  UPLOAD_SUCCESS = 'UPLOAD_SUCCESS',
  UPLOAD_FAILURE = 'UPLOAD_FAILURE'
}